+++ packages/tideline/js/tidelinedata.js
@@ -665,6 +665,37 @@ TidelineData.prototype.sortPumpSettingsParameters = function sortPumpSettingsPar
   });
 };
 
+
+TidelineData.prototype.deduplicateBoluses = function deduplicateBoluses() {
+  if (!Array.isArray(this.data)) {
+    return;
+  }
+
+  // group by time
+  const grpByTime = _.groupBy(
+    _.filter(this.data, { type: "bolus" }),
+    "normalTime"
+  );
+  const toBeRemoved = [];
+  _.forEach(grpByTime, (value) => {
+    if (value.length > 1) {
+      let goodBolus = null;
+      // Search the bolus we want to keep
+      for (const bolus of value) {
+        if (goodBolus === null) goodBolus = bolus;
+        else if (goodBolus.normal < bolus.normal) goodBolus = bolus;
+      }
+      // Add the ids we no longer want
+      for (const bolus of value) {
+        if (bolus.id !== goodBolus.id) {
+          toBeRemoved.push(bolus.id);
+        }
+      }
+    }
+  });
+  this.data = this.data.filter((d) => !toBeRemoved.includes(d.id));
+};
+
 TidelineData.prototype.deduplicatePhysicalActivities = function deduplicatePhysicalActivities() {
   const physicalActivity = this.grouped.physicalActivity;
   if (!Array.isArray(physicalActivity)) {
@@ -1100,6 +1131,11 @@ TidelineData.prototype.addData = async function addData(newData) {
   if (this.data.length < nDateBeforeRemoveDuplicates) {
     this.log.info(`${nDateBeforeRemoveDuplicates - this.data.length} duplicates data removed`);
   }
+
+  startTimer("deduplicateBoluses");
+  this.deduplicateBoluses();
+  endTimer("deduplicateBoluses");
+
   // Initial sort
   this.data.sort((a, b) => a.epoch - b.epoch);
   endTimer("Concatenate & uniq & sort");